% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auc}
\alias{auc}
\title{AUC Value from vector of predictions and labels}
\usage{
auc(preds, labels)
}
\arguments{
\item{preds}{Vector of predictions (numeric)
If the classifier returns a matrix or dataframe,
we should pass only the column relative to the predition of the label 1.
Predictions should be in the form of probabilities of belonging to the class 1}

\item{labels}{Vector of true labels (Y variable) (numeric)}
}
\value{
AUC value of the classifier underlying the predictions passed as argument
}
\description{
Works for predictions made out of classifiers.
It gives backe the Area Under the Curve (AUC) value
}
