% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cond_num}
\alias{cond_num}
\title{Evaluate the Condition Number of a matrix, allowing to discard variables}
\usage{
cond_num(dataset, vars_to_discard = NULL)
}
\arguments{
\item{dataset}{Matrice (o anche dataframe) delle variabili X}

\item{vars_to_discard}{lista con nomi delle variabili da non considerare nel calcolo del condition number}
}
\value{
Print the condition number and return it as a numeric value
}
\description{
Calcolo il condition number di una matrice di dati.
 Pensata per capire se c'è collinearità nella matrice X (correlazioni lineari alte tra gruppi di variabili).
 Per ottenere condition number reliable devo mettere le variabili sulla stessa scala. SCRIVERE PERCHÈ
 Quindi la funzione standardizza le variabili, e toglie le variabili contenute nella lista \code{vars_to_discard},
 poi calcola il condition number della matrice risultante.
 Esempi di variabili da scartare per calcolare il condition number:
We should discard all the non numeric (character) variables (or encode them using the model matrix),
discard also the response variable, and the intercept (da NaN se faccio lo scale)"
}
