% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Confusion Matrix for our classifier}
\usage{
confusion.matrix(preds, labels, threshold = 0.5)
}
\arguments{
\item{preds}{Vector of predictions (numeric)
If the classifier returns a matrix or dataframe,
we should pass only the column relative to the predition of the label 1.
Predictions should be in the form of probabilities of belonging to the class 1}

\item{labels}{Vector of true labels (Y variable) (numeric)}

\item{threshold}{Number in the range [0,1], the values in preds higher than threshold are considered predicted in class 1,
the others belong to class 0}
}
\value{
Confusion Matrix: a dataframe with the Two entry confusion matrix and a column of class errors
}
\description{
Giving the predictions (as vector of probabilities of being class 1),
and the threshold value, the function returns the Confusion Matrix
with the chosen threshold value to assign units to class 1 or 0 accordingly.
Works only for binary classifiers.
}
