% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{media_ponderata}
\alias{media_ponderata}
\title{Media Ponderata per voti di esami universitari, rispetto ai crediti}
\usage{
media_ponderata(b, c)
}
\arguments{
\item{b}{Matrice dei voti ( righe -->studenti, colonne --> esami)}

\item{c}{Vettore dei crediti di ogni esame (stesso ordine delle colonne di b)}
}
\value{
Vettore delle medie ponderate, \code{length(media_ponderata)=nrows(b)}
}
\description{
Funzione a cui passo la matrice dei voti degli esami, e un vettore
con il numero di crediti per ogni esame. Mi restituisce un vettore delle
medie voto ponderate (una per ogni studente: numero di righe della matrice \code{b})
}
