% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_test}
\alias{normalize_test}
\title{Normalize the Test Dataset, based on Train set statistics}
\usage{
normalize_test(X_test, stats_train)
}
\arguments{
\item{X_test}{Dataframe delle variabili X del Test Set}

\item{stats_train}{Lista con variabili chiamate max e min. Ognuna di queste è una lista con Nome_Variabile: Valore}
}
\value{
Un Dataframe che è il dataframe X_test con tutte le variabili normalizzate
}
\description{
Lo faccio per non avere data leakage tra Train e Test.
Normalizzo le variabili, quindi le metto tutte nel range [0,1] (del Train set),
quindi è possibile che ci siano anche valori minori di zero e maggiori di 1 (dato che lo applico sul test set)
}
